fix: 컴포넌트, 유틸 분리리
