+++ app/components/toast/Toast.module.scss
@@ -7,6 +7,7 @@
   left: 0;
   width: 100vw;
   height: 100vh;
+  z-index: 1001;
   @include flex();
 
   .toast {
